Moved login dialog to its own component
